Update Microservice A.py
I implemented additional conditional logic in the while listening loop. Now, the microservice will trigger each of it's functions when receiving certain text strings from the main program via zeromq, receiving the string "import" triggers a bank retrieval by name, receiving the string "export" triggers saving a bank to storage. Doing this also required some additional request and response logic to be added, so a confirmation is returned after the first command is received, then the main program sends the desired name of the bank.

I removed the blanket try/except block in the while listening loop, and I replaced it with more specific, localized try and except blocks. One try/except was added in the load storage method: when opening the file for read, if the file is corrupted, simply return an empty dictionary.

The second try/except, is in the export logic: try saving to storage, if there's an error, send a failure message to the main program to preserve the REQ and RES handshake in zeromq, set current key to None, and move to the next loop iteration for continued listening.